requirements.txt
# Article Deduplication System - Requirements

# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0

# Configuration Management
python-dotenv>=1.0.0

# AI Provider Dependencies
openai>=1.0.0

# Database and File Handling
sqlite3
openpyxl>=3.1.0
xlrd>=2.0.1

# Machine Learning and NLP
scikit-learn>=1.3.0
requests>=2.31.0

# Utilities
pathlib2
logging

# Uncomment if using Hugging Face transformers locally
# transformers>=4.30.0
# torch>=2.0.0
# sentence-transformers>=2.2.0

# Development Dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0